requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
requests==2.31.0
fastapi-utils==0.6.0
nest-asyncio==1.6.0
asyncio==3.4.3
PyPDF2==3.0.1
google-generativeai
requests
pillow
python-dotenv
PyMuPDF